etu_django_frame/django/4.2.2/requirements/requirement.txt
Django==4.2.2
celery==5.3.1
coreapi==2.3.3
django-cas-ng==4.3.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.1.0
django-filter==23.2
django-redis-cache==3.0.1
django-redis-sessions==0.6.2
djangorestframework==3.14.0
django-simpleui==2023.3.1
drf-yasg==1.21.6
ipython==8.14.0
kombu==5.3.1
mysqlclient==2.2.0
redis==3.5.3
requests==2.31.0
supervisor==4.2.5
uWSGI==2.0.21
vine==5.0.0
etu-django-mcmt==1.0.1
